requirements.txt
pandas
selenium 
pytest 
requests 
allure-pytest 